[src] Add vector quantization module
